+++ src/collective/contact/core/browser/static/forms.js
@@ -4,14 +4,14 @@ contactswidget = {};
 contactswidget.update_person_title = function(){
 	var gender = $(this).val();
 	var person_title = $('#form-widgets-person_title').val();
-	if (person_title == '' || person_title == 'M.' || person_title == 'Mme') {
+	if (person_title === '' || person_title == 'M.' || person_title == 'Mme') {
 		if (gender == 'M') {
 			$('#form-widgets-person_title').val("M.");
 		} else {
 			$('#form-widgets-person_title').val("Mme");
 		}
 	}
-}
+};
 
 /* Replace spaces with underscores, removes accents and other special characters */
 contactswidget.normalize_string = function(s) {
@@ -33,26 +33,26 @@ contactswidget.normalize_string = function(s) {
 	s = s.toLowerCase();
 	for (var r in rules) s = s.replace(rules[r], r);
 	return s;
-}
+};
 
 contactswidget.serialize_form = function(form) {
-   viewArr = $(form).serializeArray(),
-   view = {};
+   var viewArr = $(form).serializeArray();
+   var view = {};
    for (var i in viewArr) {
      view[viewArr[i].name] = viewArr[i].value;
    }
    return view;
- }
+ };
 
 /* Update the token which matches the input*/
 contactswidget.update_token = function(){
 	var name = $(this).val();
 	var token = $(this).closest('tr').find('input[id$="-widgets-token"]');
-	if (token.val() == '') {
+	if (token.val() === '') {
 		token_value = contactswidget.normalize_string(name);
 		token.val(token_value);
 	}
-}
+};
 
 /* Manage all stuff for the directory edit form */
 contactswidget.manage_directory = function(){
@@ -64,14 +64,19 @@ contactswidget.manage_directory = function(){
 	$('.portaltype-directory .datagridwidget-table-view thead').hide();
 	// update tokens if necessary
 	$('input[id$="-widgets-name"]').blur(contactswidget.update_token);
-}
+};
 
 /* Hide use_parent_address field if address field is empty */
 contactswidget.manage_hide_use_parent_address = function(){
-	if ($("#address").length == 0) {
-		$("#formfield-form-widgets-IContactDetails-use_parent_address").hide();
-	}
-}
+    if ($("#form-widgets-IContactDetails-parent_address").text().trim().length === 0) {
+        if($('#formfield-form-widgets-position').length === 0){
+            /* except on held position form because, there,
+             * actual parent address can change during edition
+             */
+            $("#formfield-form-widgets-IContactDetails-use_parent_address").hide();
+        }
+    }
+};
 
 contactswidget.get_selected_contact = function(form, field_id) {
   var view = contactswidget.serialize_form(form);
@@ -83,7 +88,7 @@ contactswidget.get_selected_contact = function(form, field_id) {
   var title = input.siblings('.label').find('a').first().text();
   var path = '/' + token.split('/').slice(2).join('/');
   return {token: token, title: title, path: path};
-}
+};
 
 contactswidget.setup_relation_dependency = function(master_field, slave_field, relation){
     /* slave field vocabulary is restricted on contacts who have a relation 'relation'
@@ -95,12 +100,12 @@ contactswidget.setup_relation_dependency = function(master_field, slave_field, r
         var form = input.parents('form').first();
         var selected = contactswidget.get_selected_contact(form, master_field);
         if(selected === undefined){
-            return
+            return;
         }
         /* set new relation search parameter on slave field */
         var relations = {};
         relations['relations.' + relation + ':record'] = selected.token;
-        var slave_field_query = $('#' + slave_field.replace(/\./g, '-') + '-widgets-query')
+        var slave_field_query = $('#' + slave_field.replace(/\./g, '-') + '-widgets-query');
         slave_field_query.setOptions({extraParams: relations}).flushCache();
 
         /* change create link so that master field selection is selected by default */
@@ -115,9 +120,9 @@ contactswidget.setup_relation_dependency = function(master_field, slave_field, r
             }
             add_link.attr('href', new_url);
             add_link.data('pbo').src = new_url;
-            add_link.text(add_link.orig_text + ' (' + selected.title + ')')
+            add_link.text(add_link.orig_text + ' (' + selected.title + ')');
         }
-    }
+    };
 
     var selector = '#' + master_field.replace(/\./g, '-') + '-input-fields input';
 
@@ -128,15 +133,20 @@ contactswidget.setup_relation_dependency = function(master_field, slave_field, r
     $(document).ready(function(){
         $('body').find(selector).each(function(){
             apply_relation_dependency($(this));
-        })
+        });
     });
-}
+};
 
 $(document).ready(function(){
     $(document).delegate('#formfield-form-widgets-gender input', 'change',
                          contactswidget.update_person_title);
     contactswidget.manage_directory();
     contactswidget.manage_hide_use_parent_address();
+
+    jQuery(document).bind('loadInsideOverlay',
+            function(e, pbajax, responseText, errorText, api){
+        contactswidget.manage_hide_use_parent_address();
+    });
 
     $('.contactoverlay').prepOverlay({
       subtype: 'ajax',