+++ src/collective/contact/core/browser/utils.py
@@ -9,7 +9,11 @@ from plone.schemaeditor.utils import non_fieldset_fields
 from plone.dexterity.interfaces import IDexterityFTI
 from plone.app.dexterity.behaviors.metadata import IBasic
 
+from collective.contact.core.behaviors import IBirthday
 from collective.contact.core.behaviors import IContactDetails
+
+
+IGNORED_BEHAVIORS = [IContactDetails, IBasic, IBirthday]
 
 
 def date_to_DateTime(date):
@@ -31,7 +35,7 @@ def get_ttw_fields(obj):
 
     for behavior_id in fti.behaviors:
         behavior = getUtility(IBehavior, behavior_id).interface
-        if behavior in (IContactDetails, IBasic) or not IFormFieldProvider.providedBy(behavior):
+        if behavior in IGNORED_BEHAVIORS or not IFormFieldProvider.providedBy(behavior):
             continue
 
         try:
@@ -44,3 +48,12 @@ def get_ttw_fields(obj):
             pass
 
     return new_fields
+
+
+def get_valid_url(url):
+    """Returns valid url (i.e. an url which starts with http or https)
+    """
+    if url and not url.startswith('http'):
+        return "http://{}".format(url)
+    else:
+        return url